SimpleVectorSearch_RAG/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6

# Vertex AI (Gemini + Vector Search + Embeddings)
google-cloud-aiplatform==1.80.0
#google-cloud-storage==3.1.1
pypdf==5.3.0

google-cloud-storage>=2.19.0,<3.0.0

# Firestore — scalable chunk metadata store
# Replaces loading the entire JSONL into RAM; fetches only top-K docs per query.
google-cloud-firestore>=2.19.0

# Streamlit UI
streamlit>=1.41.0
requests>=2.32.0
